change getcampOpens to be in campaigns.go
