+++ .gitignore
@@ -32,4 +32,8 @@ static
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 
+# Goland & .vcscode
+*.vscode
+*.idea
+
 .env*
